+++ UI/src/plugins/storybook.js
@@ -1,8 +1,5 @@
-import 'core-js'
 import Vue from 'vue'
 import VueConfirmDialog from 'vue-confirm-dialog'
-
-// Vuetify
 import Vuetify from 'vuetify'
 import 'vuetify/dist/vuetify.min.css'
